+++ energieha/src/web/__init__.py
@@ -1,0 +1,1 @@
+"""EnergieHA web interface package."""